+++ routes/jwt.mjs
@@ -83,4 +83,4 @@ router.get(
   },
 );
 
-export default router.routes();
+export default router;