+++ routes/rides.mjs
@@ -294,4 +294,4 @@ router.post(
   },
 );
 
-export default router.routes();
+export default router;