+++ routes/users.mjs
@@ -244,4 +244,4 @@ router.post(
   },
 );
 
-export default router.routes();
+export default router;