+++ routes/campuses/cars.mjs
@@ -35,4 +35,4 @@ router.get(
   },
 );
 
-export default router.routes();
+export default router;