+++ routes/feedback.mjs
@@ -43,4 +43,4 @@ router.post(
   },
 );
 
-export default router.routes();
+export default router;