+++ routes/logs.mjs
@@ -38,4 +38,4 @@ router.get(
   },
 );
 
-export default router.routes();
+export default router;