+++ routes/phones.mjs
@@ -73,4 +73,4 @@ const router = generateCRUD(Phone, {
   },
 });
 
-export default router.routes();
+export default router;