+++ routes/pois.mjs
@@ -66,4 +66,4 @@ const router = generateCRUD(Poi, {
   },
 });
 
-export default router.routes();
+export default router;