+++ routes/campuses/drivers-positions.mjs
@@ -21,4 +21,4 @@ router.get(
   },
 );
 
-export default router.routes();
+export default router;