+++ routes/users.mjs
@@ -2,7 +2,8 @@ import generateCRUD from '../helpers/abstract-route';
 import User from '../models/user';
 import NotificationDevice from '../models/notification-device';
 import {
-  CAN_CREATE_USER, CAN_EDIT_SELF_USER_NAME, CAN_EDIT_SELF_USER_PASSWORD,
+  CAN_CREATE_USER,
+  CAN_EDIT_SELF_USER_NAME, CAN_EDIT_SELF_USER_PASSWORD, CAN_EDIT_SELF_USER_SENSITIVE_DATA,
   CAN_SEND_CREATION_TOKEN,
   CAN_EDIT_USER,
   CAN_EDIT_USER_SENSITIVE_DATA,
@@ -147,16 +148,23 @@ const router = generateCRUD(User, {
         userBody.password = body.password;
       }
 
-      if (ctx.may(CAN_EDIT_SELF_USER_NAME) && body.name) {
-        userBody.name = body.name;
+      if (ctx.may(CAN_EDIT_SELF_USER_SENSITIVE_DATA)) {
+        if (body.phone) {
+          userBody.phone = body.phone;
+        }
+        if (body.email) {
+          userBody.email = body.email;
+          userBody.email_token = body.email_token;
+        }
       }
 
-      if (ctx.may(CAN_EDIT_SELF_USER_NAME) && body.firstname) {
-        userBody.firstname = body.firstname;
-      }
-
-      if (ctx.may(CAN_EDIT_SELF_USER_NAME) && body.lastname) {
-        userBody.lastname = body.lastname;
+      if (ctx.may(CAN_EDIT_SELF_USER_NAME)) {
+        if (body.firstname) {
+          userBody.firstname = body.firstname;
+        }
+        if (body.lastname) {
+          userBody.lastname = body.lastname;
+        }
       }
 
       const user = await User.findById(id);
@@ -173,7 +181,10 @@ const router = generateCRUD(User, {
         );
       }
 
-      if (!ctx.may(CAN_EDIT_USER_SENSITIVE_DATA)) {
+      if (
+        !ctx.may(CAN_EDIT_USER_SENSITIVE_DATA)
+        && !ctx.may(CAN_EDIT_SELF_USER_SENSITIVE_DATA)
+      ) {
         delete userBody.email_confirmed;
         if (userBody.phone) {
           delete userBody.phone.canonical;
@@ -182,7 +193,6 @@ const router = generateCRUD(User, {
       }
 
       user.set(userBody);
-
       if (!user.phone.confirmed && userBody.phone && userBody.phone.token) {
         await user.confirmPhone(body.phone.token);
       }
@@ -196,7 +206,6 @@ const router = generateCRUD(User, {
         'You\'re not authorized to update this user',
       );
       ctx.body = await user.save();
-
       if ((ctx.headers[X_SEND_TOKEN] && ctx.headers[X_SEND_TOKEN] !== 'false')) {
         const toSend = ctx.headers[X_SEND_TOKEN].split(',');
         if (toSend.includes('email') && !user.email_confirmed) {