+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## Current
 ### Features
 ### Fix
+* Throw error if ride has the same departure and arrival POI
 ## Version 1.4.5
 ### Features
 * Added connection status management for drivers through sockets