+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 ## Current
 ### Features
 * Ensure that MX entry exists for an email address
+* Add a limit to email size.
 ### Fix
+* Throw error if ride has the same departure and arrival POI
 ## Version 1.4.5
 ### Features
 * Added connection status management for drivers through sockets