+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 ## Current
 ### Features
+* Ensure that MX entry exists for an email address
 * Add a limit to email size.
 ### Fix
 * Throw error if ride has the same departure and arrival POI