+++ CHANGELOG.md
@@ -1,9 +1,21 @@
 # Changelog
 ## Current
 ### Features
+* Enforce allowed methods error
+* Enforce CORS allowed origin
+* Add few HTTP headers thanks to helmet
+* Ensure that MX entry exists for an email address
+* Add a limit to email size.
 ### Fix
+* Normalize email in find
+* Ensure that email is compliant with Regex
+* Check filter for phone listing
+* Don't let admin remove a user
+* Ensure that car is linked to the current selected campus
+* Clean POI authorizations 
 * Throw error if ride has the same departure and arrival POI
 * Added rate limit model & handles it for password attempt
+* Fixed rights for user update
 ## Version 1.4.5
 ### Features
 * Added connection status management for drivers through sockets