+++ CHANGELOG.md
@@ -19,6 +19,9 @@
 * Added required to coordinates in campus model
 * Added rights to drivers-positions.mjs
 * Scoped CRUD rights to campus
+* CSV import duplication error message
+* id to _id conversion
+* Refs for batch phones route
 ## Version 1.4.5
 ### Features
 * Added connection status management for drivers through sockets